One more personography example and placeography example
Added a new person, John Doe, with personagraphy info in the header. Also added a new place, Boston, wiwth personography info in the header. -ography data for both are included in note element instead of the series of elements in the original example -ography data. We're curious how this will display in TAPAS and will be testing soon.