Clarifications
Other than some clarification, I didn't change the fact that the note that I'm testing with an image inside it is contained within a paragraph. There are two differences between this (and the previous) iteration of this file and the earlier iteration that successfully displayed an image inside a note. The first is that the note came after a close paragraph tag. The second is that there is no head or figDesc inside the figure element.